refactor serverFetchQueries to call apiProxy$ directly
