+++ src/plugins/requestAuthPlugin.js
@@ -39,7 +39,7 @@ const handleLogout = request => {
 
 function getAuthType(request) {
 	const memberCookie = getMemberCookieName(request.server);
-	const allowedAuthTypes = [memberCookie, 'oauth_token', '__internal_oauth_token'];
+	const allowedAuthTypes = [memberCookie, 'oauth_token'];
 	// search for a request.state cookie name that matches an allowed auth type
 	return allowedAuthTypes.reduce(
 		(authType, allowedType) => authType || request.state[allowedType] && allowedType,
