+++ src/util/apiUtils.js
@@ -212,15 +212,16 @@ export const apiResponseToQueryResponse = query => ({ value, meta }) => ({
 	}
 });
 
-export function getAuthHeaders({ state }) {
+export function getAuthHeaders(request) {
 	// internal server requests may set non-encoded token cookie __internal_oauth_token
-	const oauth_token = state.oauth_token || state.__internal_oauth_token;
-	if (!state.MEETUP_MEMBER && oauth_token) {
+	const oauth_token = request.state.oauth_token ||
+		request.plugins.requestAuth && request.plugins.requestAuth.oauth_token;  // for internal requests
+	if (!request.state.MEETUP_MEMBER && oauth_token) {
 		return {
 			authorization: `Bearer ${oauth_token}`,
 		};
 	}
-	const cookies = { ...state };
+	const cookies = { ...request.state };
 	const csrf = uuid.v4();
 	cookies.MEETUP_CSRF = csrf;
 	cookies.MEETUP_CSRF_DEV = csrf;
@@ -259,6 +260,11 @@ export function parseRequestQueries(request) {
 		query,
 	} = request;
 	const queriesJSON = method === 'post' ? payload.queries : query.queries;
+
+	if (!queriesJSON) {
+		return null;
+	}
+
 	const validatedQueries = Joi.validate(
 		JSON.parse(queriesJSON),
 		Joi.array().items(querySchema)
