+++ src/util/authUtils.js
@@ -47,7 +47,7 @@ export const applyAuthState = (request, reply) => auth => {
 	Object.keys(authState).forEach(name => {
 		const cookieVal = authState[name];
 		// apply to request
-		request.state[`__internal_${name}`] = cookieVal.value;  // this will only be used for generating internal requests
+		request.plugins.requestAuth[name] = cookieVal.value;  // this will only be used for generating internal requests
 		// apply to response - note this special `request.authorize.reply` prop assigned onPreAuth
 		reply.state(name, cookieVal.value, cookieVal.opts);
 	});
